+++ microservices/transaction/controller.py
@@ -1,0 +1,4 @@
+from repository import fetch_transactions
+
+def get_transactions():
+    return {"data": fetch_transactions()}
