+++ microservices/transaction/router.py
@@ -1,0 +1,8 @@
+from fastapi import APIRouter
+from controller import get_transactions
+
+router = APIRouter()
+
+@router.get("/")
+def transactions():
+    return get_transactions()
